refactor(cli): migrate credentials to TypeScript
